Create practiceGagesHardCode
Hard codes values to create a ncdf file. Not particularly useful, but shows possible methods to parse through data, depending on what format the wml2 data was formatted into. Need to change the location of practiceGages to where it is saved. Trying to fill multiple points of data doesn't work in this example: see createNCDF.py for a working example